Mkrd: fixed size of lbastart and lbalength fields. Now it works on Mac OS X.

+++ tools/mkrd/rdisk.h
@@ -18,8 +18,8 @@ struct partition {
 	struct chs start;
 	unsigned char type;
 	struct chs end;
-	unsigned long lbastart;
-	unsigned long lbalength;
+	unsigned lbastart;
+	unsigned lbalength;
 };
 
 struct mbr {
